patches: Update rstudio-server-multi-version patches.
The changes have been rebased on top of version 1.4.1717 and include a
bug fix.  Previous versions would fail to look up the user's active
sessions (and thus fail to restore a previously set version of R) when
rserver did not run as the same user account as the rsession
process (e.g. root).