python: Add python-jaxlib.
This commit adds what might later become a generic bazel-build-system.
BAZEL-VENDORED-INPUTS is a procedure that uses Bazel to compute a
source code bundle for a given configuration.  This bundle is then
used as an input in the actual package where Bazel is convinced to
accept the bundle instead of fetching source code over the internet
all over again.

* guix-science/packages/python.scm (bazel-vendored-inputs): New
procedure.
(python-jaxlib/wheel, python-jaxlib): New variables.