python: Do not use "--nofetch" for Bazel builds.
"--nofetch" sounds like the right thing to do, but when it is present
Bazel is unwilling to accept repositories provided at
$NIX_BUILD_TOP/output/external.  Notably, this means that the
essential @bazel_tools would be missing.

We can drop "--nofetch" only because of the patch to Bazel in the
previous commit, which tells Bazel to unconditionally recompute hashes
for repository marker files (e.g. @bazel_tools.marker).  Without the
patch, the omission of "--nofetch" would result in repositories to be
downloaded because the checksum in the (empty) marker files doesn't
match the contents of the prepared repositories in
$NIX_BUILD_TOP/output/external.

* guix-science/packages/python.scm (python-jaxlib/wheel, tensorflow,
python-tensorflow-probability)[arguments]: Remove "--nofetch" from
"bazel run" invocation.